refactor: Setting 내 코드들 Auth 내부로 이동 (apis, services, types) (#37)
