+++ src/apis/index.ts
@@ -3,4 +3,3 @@ export * from './book';
 export * from './file';
 export * from './myLibrary';
 export * from './record';
-export * from './setting';
