+++ src/services/index.ts
@@ -2,5 +2,4 @@ export * from './auth';
 export * from './book';
 export * from './myLibrary';
 export * from './record';
-export * from './setting';
 export { default as queryClient } from './queryClient';
